refactor:分离各个bff的请求结构体与响应结构体
