+++ interface/storage/api/model/model.go
@@ -1,6 +1,5 @@
-package utils
+package model
 
-// Response 通用响应体
 type Response struct {
 	Code int         `json:"code"`
 	Msg  string      `json:"msg"`
